requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
firebase-admin==6.5.0
google-cloud-bigquery==3.25.0
cloud-sql-python-connector[pg8000]==1.12.0
SQLAlchemy==2.0.36
pyfarmhash==0.3.2
pydantic==2.9.2
